fixed overflow errors
% go test -run TestEnmarshalVarbind
./marshal_test.go:263: constant 4294967295 overflows int
./verax_test.go:159: constant 4294967295 overflows int
FAIL    _/home/sonia/go/bpl-poller/src/github.com/soniah/gosnmp [build
failed]

+++ marshal_test.go
@@ -260,7 +260,7 @@ var testsUnmarshal = []struct {
 				{
 					Name:  "1.3.6.1.2.1.2.2.1.5.3",
 					Type:  Gauge32,
-					Value: 4294967295,
+					Value: uint(4294967295),
 				},
 				{
 					Name:  "1.3.6.1.2.1.2.2.1.7.2",
